requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
openpyxl==3.0.6
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
